Add files via upload
Thanks to https://github.com/tony-tiger for providing the .raw files. this is a .wav file that Will go through all pager restaurant ID's and Pager ID's. Just put in the WAV folder on your mayhem SDcard open soundboard on your poratpack tune to ferq 467.750mhz. I have not tested this please feel free to tell me if it does not work.